Remove redundant tests
