+++ src/test/java/com/pivovarit/collectors/parallel/ParallelismThrottlingBDDTest.java
@@ -1,4 +1,4 @@
-package com.pivovarit.collectors.parallelToCollection;
+package com.pivovarit.collectors.parallel;
 
 import com.pholser.junit.quickcheck.Property;
 import com.pholser.junit.quickcheck.generator.InRange;
@@ -9,33 +9,32 @@ import org.awaitility.Awaitility;
 import org.junit.runner.RunWith;
 
 import java.time.Duration;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static com.pivovarit.collectors.ParallelCollectors.parallel;
-import static com.pivovarit.collectors.ParallelCollectors.parallelToStream;
 import static com.pivovarit.collectors.infrastructure.TestUtils.TRIALS;
 import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelay;
 import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelayGaussian;
 import static java.time.Duration.ofMillis;
-import static java.util.stream.Collectors.toCollection;
+import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author Grzegorz Piwowarek
  */
 @RunWith(JUnitQuickcheck.class)
-public class ToCollectionParallelismThrottlingBDDTest extends ExecutorAwareTest {
+public class ParallelismThrottlingBDDTest extends ExecutorAwareTest {
 
     @Property(trials = TRIALS)
-    public void shouldCollectToCollectionWithThrottledParallelism(@InRange(minInt = 20, maxInt = 100) int unitsOfWork, @InRange(minInt = 1, maxInt = 20) int parallelism) {
+    public void shouldCollectToListWithThrottledParallelism(@InRange(minInt = 20, maxInt = 100) int unitsOfWork, @InRange(minInt = 1, maxInt = 20) int parallelism) {
         // given
         TestUtils.CountingExecutor executor = new TestUtils.CountingExecutor();
 
         Stream.generate(() -> 42)
           .limit(unitsOfWork)
-          .collect(parallelToStream(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), executor, parallelism));
+          .collect(parallel(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), toList(), executor, parallelism));
 
         Awaitility.await()
           .until(() -> executor.count() == parallelism);
@@ -46,11 +45,11 @@ public class ToCollectionParallelismThrottlingBDDTest extends ExecutorAwareTest 
         // given
         executor = threadPoolExecutor(unitsOfWork);
         List<Integer> result = Stream.iterate(0, i -> i + 1).limit(20)
-          .collect(parallel(i -> returnWithDelayGaussian(i, Duration
-            .ofMillis(10)), toCollection(ArrayList::new), executor, parallelism))
+          .collect(parallel(i -> returnWithDelayGaussian(i, Duration.ofMillis(10)), toList(), executor, parallelism))
           .join();
 
         assertThat(result).isSorted();
-        executor.shutdown();
+
+        executor.shutdownNow();
     }
 }
