+++ src/test/java/com/pivovarit/collectors/parallel/ParallelismThrottlingTest.java
@@ -1,34 +1,34 @@
-package com.pivovarit.collectors.parallelToCollection;
+package com.pivovarit.collectors.parallel;
 
 import com.pivovarit.collectors.infrastructure.TestUtils;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static com.pivovarit.collectors.ParallelCollectors.parallel;
 import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelay;
 import static java.time.Duration.ofMillis;
-import static java.util.stream.Collectors.toCollection;
+import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author Grzegorz Piwowarek
  */
-class ToCollectionParallelismThrottlingTest {
+class ParallelismThrottlingTest {
 
     @Test
-    void shouldParallelizeToListAndRespectParallelizm() {
+    void shouldParallelizeToSetAndRespectParallelizm() {
         // given
         int parallelism = 2;
         TestUtils.CountingExecutor executor = new TestUtils.CountingExecutor();
 
-        CompletableFuture<ArrayList<Long>> result = Stream.generate(() -> 42L)
-          .limit(10)
-          .collect(parallel(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), toCollection(ArrayList::new), executor, parallelism));
+        CompletableFuture<List<Long>> result =
+          Stream.generate(() -> 42L)
+            .limit(10)
+            .collect(parallel(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), toList(), executor, parallelism));
 
         assertThat(result)
           .isNotCompleted()
@@ -38,16 +38,18 @@ class ToCollectionParallelismThrottlingTest {
           .until(() -> executor.count() == parallelism);
     }
 
+
     @Test
-    void shouldParallelizeToListAndRespectParallelizmMapping() {
+    void shouldParallelizeToSetAndRespectParallelizmMapping() {
         // given
         int parallelism = 2;
         TestUtils.CountingExecutor executor = new TestUtils.CountingExecutor();
 
-        CompletableFuture<ArrayList<Long>> result =
+
+        CompletableFuture<List<Long>> result =
           Stream.generate(() -> 42)
             .limit(10)
-            .collect(parallel(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), toCollection(ArrayList::new), executor, parallelism));
+            .collect(parallel(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), toList(), executor, parallelism));
 
         assertThat(result)
           .isNotCompleted()
