--- src/test/java/com/pivovarit/collectors/parallelToSet/ToSetParallelismThrottlingBDDTest.java
@@ -1,37 +1,0 @@
-package com.pivovarit.collectors.parallelToSet;
-
-import com.pholser.junit.quickcheck.Property;
-import com.pholser.junit.quickcheck.generator.InRange;
-import com.pholser.junit.quickcheck.runner.JUnitQuickcheck;
-import com.pivovarit.collectors.infrastructure.ExecutorAwareTest;
-import com.pivovarit.collectors.infrastructure.TestUtils;
-import org.awaitility.Awaitility;
-import org.junit.runner.RunWith;
-
-import java.util.stream.Stream;
-
-import static com.pivovarit.collectors.ParallelCollectors.parallel;
-import static com.pivovarit.collectors.infrastructure.TestUtils.TRIALS;
-import static com.pivovarit.collectors.infrastructure.TestUtils.returnWithDelay;
-import static java.time.Duration.ofMillis;
-import static java.util.stream.Collectors.toSet;
-
-/**
- * @author Grzegorz Piwowarek
- */
-@RunWith(JUnitQuickcheck.class)
-public class ToSetParallelismThrottlingBDDTest extends ExecutorAwareTest {
-
-    @Property(trials = TRIALS)
-    public void shouldCollectToSetWithThrottledParallelism(@InRange(minInt = 20, maxInt = 100) int unitsOfWork, @InRange(minInt = 1, maxInt = 20) int parallelism) {
-        // given
-        TestUtils.CountingExecutor executor = new TestUtils.CountingExecutor();
-
-        Stream.generate(() -> 42)
-          .limit(unitsOfWork)
-          .collect(parallel(i -> returnWithDelay(42L, ofMillis(Integer.MAX_VALUE)), toSet(), executor, parallelism));
-
-        Awaitility.await()
-          .until(() -> executor.count() == parallelism);
-    }
-}
